+++ routes/orders.rb
@@ -15,7 +15,7 @@ class BitcoinExchange < Sinatra::Base
   post "/orders" do
     # place a buy / sell order
     order = params[:order] || {}
-    attributes = { user_id: current_user.id, type: order[:type].to_sym, amount: order[:amount].to_f }
+    attributes = { user_id: current_user.id, type: order[:type].to_sym, amount: order[:amount].to_d }
     if order[:simple]
       Order.create_simple attributes
     else
